+++ tests/integration_tests/headers.rs
@@ -4,15 +4,11 @@ use crate::utils::*;
 fn it_adds_new_header() -> TestResult {
     let quartz = Quartz::preset_using_sample_endpoint()?;
 
-    let headers_add_output = quartz.cmd(&["headers", "--add", "Content-type: application/json"])?;
-    let headers_output = quartz.cmd(&["headers", "--get", "Content-type"])?;
+    let set_output = quartz.cmd(&["header", "--set", "Content-type: application/json"])?;
+    let output = quartz.cmd(&["header", "--get", "Content-type"])?;
 
-    assert!(
-        headers_add_output.status.success(),
-        "{}",
-        headers_add_output.stderr
-    );
-    assert_eq!(headers_output.stdout.trim(), "application/json");
+    assert!(set_output.status.success(), "{}", set_output.stderr);
+    assert_eq!(output.stdout.trim(), "application/json");
 
     Ok(())
 }
@@ -28,15 +24,15 @@ fn it_adds_multiple_new_header() -> TestResult {
     ];
 
     let headers_add_output = quartz.cmd(&[
-        "headers",
-        "--add",
+        "header",
+        "--set",
         sample_headers[0],
-        "--add",
+        "--set",
         sample_headers[1],
-        "--add",
+        "--set",
         sample_headers[2],
     ])?;
-    let headers_output = quartz.cmd(&["headers", "--list"])?;
+    let output = quartz.cmd(&["header", "--list"])?;
 
     assert!(
         headers_add_output.status.success(),
@@ -45,7 +41,7 @@ fn it_adds_multiple_new_header() -> TestResult {
     );
 
     for header in sample_headers {
-        assert!(headers_output.stdout.contains(header));
+        assert!(output.stdout.contains(header));
     }
 
     Ok(())
@@ -56,30 +52,24 @@ fn it_overwrites_existing_headers() -> TestResult {
     let quartz = Quartz::preset_using_sample_endpoint()?;
 
     quartz.cmd(&[
-        "headers",
-        "--add",
+        "header",
+        "--set",
         "Content-type: application/json",
-        "--add",
+        "--set",
         "Accept: application/json",
     ])?;
 
-    let headers_edit_output = quartz.cmd(&["headers", "--add", "Content-type: plain/text"])?;
-    let headers_output = quartz.cmd(&["headers", "--list"])?;
+    let edit_output = quartz.cmd(&["header", "--set", "Content-type: plain/text"])?;
+    let output = quartz.cmd(&["header", "--list"])?;
+
+    assert!(edit_output.status.success(), "{}", edit_output.stdout);
 
     assert!(
-        headers_edit_output.status.success(),
-        "{}",
-        headers_edit_output.stdout
-    );
-
-    assert!(
-        !headers_output
-            .stdout
-            .contains("Content-type: application/json"),
+        !output.stdout.contains("Content-type: application/json"),
         "old value found"
     );
     assert!(
-        headers_output.stdout.contains("Content-type: plain/text"),
+        output.stdout.contains("Content-type: plain/text"),
         "new value was not saved"
     );
 
@@ -91,14 +81,14 @@ fn it_removes_header_by_key() -> TestResult {
     let quartz = Quartz::preset_using_sample_endpoint()?;
 
     quartz.cmd(&[
-        "headers",
-        "--add",
+        "header",
+        "--set",
         "Content-type: application/json",
-        "--add",
+        "--set",
         "Accept: form",
     ])?;
 
-    let remove_output = quartz.cmd(&["headers", "--remove", "Content-type", "--list"])?;
+    let remove_output = quartz.cmd(&["header", "--remove", "Content-type", "--list"])?;
     assert!(remove_output.status.success(), "{}", remove_output.stderr);
     assert!(
         !remove_output.stdout.contains("Content-type"),
@@ -116,21 +106,21 @@ fn it_removes_header_by_key() -> TestResult {
 fn it_does_not_allow_invalid_header_format() -> TestResult {
     let quartz = Quartz::preset_using_sample_endpoint()?;
 
-    let headers_output = quartz.cmd(&["headers", "--add", "Content-type"])?;
+    let output = quartz.cmd(&["header", "--set", "Content-type"])?;
     assert!(
-        !headers_output.status.success(),
+        !output.status.success(),
         "allowed header without value separation"
     );
 
-    let headers_output = quartz.cmd(&["headers", "--add", "Content-type = application/json"])?;
+    let output = quartz.cmd(&["header", "--set", "Content-type = application/json"])?;
     assert!(
-        !headers_output.status.success(),
+        !output.status.success(),
         "allowed header with incorrect key-value separation"
     );
 
-    let headers_output = quartz.cmd(&["headers", "--add", "Content-type:application/json"])?;
+    let output = quartz.cmd(&["headers", "--set", "Content-type:application/json"])?;
     assert!(
-        !headers_output.status.success(),
+        !output.status.success(),
         "allowed header without proper spacing between key and value"
     );
 
@@ -142,9 +132,9 @@ fn compatible_with_apply_context_option() -> TestResult {
     let quartz = Quartz::preset_using_sample_endpoint()?;
 
     quartz.cmd(&["var", "--set", "contentType=application/json"])?;
-    quartz.cmd(&["headers", "--add", "Content-type: {{contentType}}"])?;
+    quartz.cmd(&["header", "--set", "Content-type: {{contentType}}"])?;
 
-    let output = quartz.cmd(&["--apply-context", "headers", "--get", "Content-type"])?;
+    let output = quartz.cmd(&["--apply-context", "header", "--get", "Content-type"])?;
 
     assert!(output.status.success(), "{}", output.stderr);
     assert_eq!(output.stdout.trim(), "application/json");